cnss: Disable wlan irq when PCIe link down is detected
Interrupt Storm is observed on the APPS side, when PCIe link down is
simulated, as FW keeps generating interrupts to host, as it's not aware
of the PCIe link down.

When PCIe link down is identified, wlan platform driver schedules work
to start recovery. During Recovery we cut down power to Rome, which
stops the interrupt storm. During the window where platform driver
schedule work and the work actually gets schedules, interrupt storm
happened and the APPS triggered watchdog.

Change-Id: Ica3c32007bc9b8acce79c5bb2175790ff27180d0
CRs-Fixed: 967956
Signed-off-by: Komal Kumar <kseelam@codeaurora.org>
Signed-off-by: engstk <eng.stk@sapo.pt>