dts: msm8994-pm: Add additional interrupts to MPM ignore list
Add more IRQs which are hit often but are not
expected to interrupt the MPM controlled shutdown.

With these IRQs in the ignore list we can expect
MPM to vote for a modem shutdown more often,
which saves power.

Change-Id: I43fa13a40355ec90035e3764f34710f66f445844
Co-authored-by: Jake Weinstein <xboxlover360@gmail.com>
Signed-off-by: Alex Naidis <alex.naidis@linux.com>
Reviewed-on: https://gerrit.nextbit.com:8443/21176
Reviewed-by: Jake Weinstein <xboxlover360@gmail.com>
Tested-by: Jake Weinstein <xboxlover360@gmail.com>
Reviewed-by: William McVicker <william@nextbit.com>

@aviraxp note:

I highly suspect this commit as it ignores some very incremental
interrupts and some of them are no-op. But anyway gove it a test.

Signed-off-by: Wang Han <416810799@qq.com>