qcacld-2.0: Enable Wifi channel bonding mode unconditionally
Wifi channel bonding is known to improve the internet speed. But
it depends on userspace configuration, so we need to tweak it with
something like Magisk.

Avoid this by hardcoding the min/max and default limits in wlan API.
Note that it is a _HACK_, so only use this when you really, really
need it.

Signed-off-by: Wang Han <416810799@qq.com>