core_ctl: Remove online delay tunable
This feature is buggy. As a result, not_preferred tunable is not
working because of the bogus CPU selection logic. It is likely to
be caused by the is_available check.

As the actual derp is hard to find, just remove the whole tunable
as it is already zero in previous commits.

Signed-off-by: Wang Han <416810799@qq.com>