qpnp-smbcharger: Accept unknown power supply type on boot
Previously if power supply type is unknown, healthd will fail to
phrase the properties. But as we have fixed the issue in healthd,
it is necessary to do this hack.

Furthermore, if the power supply type is USB, we would possibly
end up with slow charging rate because the charging rate limit is
set to 900mA.

This reverts commit dd8065701a3af701aedbbc742baaf5434734b96f.

Signed-off by: Wang Han <416810799@qq.com>