[PATCH] floppy: Do not copy a kernel pointer to user memory in FDGETPRM ioctl CVE-2018-7755
FromBrian Belleville
Date Wed, 7 Mar 2018 16:02:45 -0800

The final field of a floppy_struct is the field name, which is a
pointer to a string in kernel memory. The kernel pointer should not be
copied to user memory. The FDGETPRM ioctl copies a floppy_struct to
user memory, including the name field. This pointer cannot be used
by the user, and it will leak a kernel address to user-space, which
will reveal the location of kernel code and data and undermine KASLR
protection. Instead, copy the floppy_struct except for the name
field.

Signed-off-by: Brian Belleville <bbellevi@uci.edu>